Fix licensing and attribution
The previous licensing information was incorrect and the person who
created the LICENSE file did not have permission to license the code
they used under that license. This corrected licensing is the ONLY
licensing that is actually valid, so you cannot use previous versions
under the previously stated, invalid, license.